+++ inst/doc/archeoViz-fr.R
@@ -1,10 +1,10 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  # déterminez le répertoire de travail dans votre shiny server:
 #  setwd(dir = "/some/path/")
 #  # téléchargez le package:
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 #  # décompression:
 #  unzip(zipfile = "archeoviz.zip")
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(objects.df = NULL,   # data.frame pour les objets
 #            refits.df = NULL,    # data.frame optionnel pour les remontages
 #            timeline.df = NULL,  # data.frame optionnel pour la chronologie des fouilles
@@ -24,18 +24,19 @@ knitr::opts_chunk$set(
 #            lang = "fr"          # langue de l'interface ("de": Allemand, "en": Anglais, "fr": Français, "it": Italien "pt": Portugais, "es": Espagnol)
 #            set.theme = "cosmo") # thème graphique de l'interface Shiny
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz()
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(objects.df = NULL,  # data.frame pour les objets
 #            refits.df = NULL,   # data.frame optionnel pour les remontages
 #            timeline.df = NULL) # data.frame optionnel pour la chronologie
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(objects.df=NULL, refits.df=NULL, timeline.df=NULL,
 #            title=NULL, home.text=NULL, lang="en", set.theme="cosmo",
-#            square.size = 100, reverse.axis.values = NULL, reverse.square.names = NULL,
+#            square.size = 100, rotation = 0, grid.orientation = NULL,
+#            reverse.axis.values = NULL, reverse.square.names = NULL,
 #            add.x.square.labels = NULL, add.y.square.labels = NULL,
 #            class.variable = NULL, class.values = NULL,
 #            default.group = "by.layer", location.mode = NULL,
@@ -45,16 +46,16 @@ knitr::opts_chunk$set(
 #            sectionX.x.val = NULL, sectionX.y.val = NULL, sectionX.refits = NULL,
 #            sectionY.x.val = NULL, sectionY.y.val = NULL, sectionY.refits = NULL,
 #            camera.center = c(0, 0, 0), camera.eye = c(1.25, 1.25, 1.25),
-#            run.plots = FALSE, html.export = TRUE, table.export =TRUE
+#            run.plots = FALSE, html.export = TRUE, table.export = TRUE
 #            )
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(square.size = 100,
 #            reverse.axis.values = NULL, reverse.square.names = NULL,
 #            add.x.square.labels = NULL, add.y.square.labels = NULL
 #            )
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(class.variable = NULL, class.values = NULL,
 #  		  default.group = "by.layer", location.mode = NULL,
 #  		  map.z.val = NULL, map.density = "no", map.refits = NULL,
@@ -64,6 +65,6 @@ knitr::opts_chunk$set(
 #  		  camera.center = NULL, camera.eye = NULL
 #            )
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  archeoViz(run.plots = FALSE)